Rename e-annotation-wrapper to e-image-wrapper, move utils
