+++ packages/core/src/components/grading/GradingAnswer.less
@@ -114,7 +114,7 @@
   padding: 10px;
   word-break: break-word;
   background: @grey-010;
-  .e-annotation-wrapper {
+  .e-image-wrapper {
     line-height: 0;
     sup {
       white-space: nowrap;
@@ -128,7 +128,7 @@
   }
 
   img,
-  .e-annotation-wrapper .e-annotation {
+  .e-image-wrapper .e-annotation {
     cursor: crosshair;
     -moz-user-drag: -moz-none;
     -khtml-user-drag: none;
