+++ packages/core/src/components/grading/editAnnotations.ts
@@ -122,7 +122,7 @@ export function hasTextSelectedInAnswerText(): boolean {
 export function selectionHasNothingToUnderline(range: Range) {
   const contents = range.cloneContents()
   const hasImages = Array.from(contents.childNodes).some(
-    x => x instanceof Element && (x.classList.contains('e-annotation-wrapper') || x.tagName === 'IMG')
+    x => x instanceof Element && (x.classList.contains('e-image-wrapper') || x.tagName === 'IMG')
   )
   return contents.textContent?.length === 0 && !hasImages
 }
@@ -227,7 +227,7 @@ export function imageAnnotationMouseDownInfo(
 ): NewImageAnnotation {
   const targetAnswerText = image.closest('.e-grading-answer')!
   const attachmentIndex = Array.from(targetAnswerText.querySelectorAll('img')).findIndex(img => img === image)
-  const attachmentWrapper = targetAnswerText.querySelectorAll('.e-annotation-wrapper').item(attachmentIndex)
+  const attachmentWrapper = targetAnswerText.querySelectorAll('.e-image-wrapper').item(attachmentIndex)
   const bbox = attachmentWrapper.getBoundingClientRect()
   const clientX = e.clientX
   const startX = clamp((clientX - bbox.left) / bbox.width)
